rename: 카멜케이스 컨벤션으로 폴더명, 파일명 이름 수정 (#122)
